+++ MordoEngine/src/Core/Managers/Implementation/ShaderManager.cpp
@@ -21,6 +21,12 @@ void Manager<Shader>::Init() {
 			FileSystem::getPath("res/shaders/skyBox.vs"),
 			FileSystem::getPath("res/shaders/skyBox.fs"))
 	);
+
+	resources.emplace("light_cube",
+		Shader(
+			FileSystem::getPath("res/shaders/light_cube.vs"),
+			FileSystem::getPath("res/shaders/light_cube.fs"))
+	);
 }
 
 template class Manager<Shader>;