+++ MordoEngine/src/Lighting/PointLight/PointLightData.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <glm/vec3.hpp>
+
+struct PointLightData {
+	glm::vec3 position;
+	glm::vec3 ambient;
+	glm::vec3 diffuse;
+	glm::vec3 specular;
+	float constant;
+	float linear;
+	float quadratic;
+};